Fix import errors and add index.js to features and move files
